// models/AccountQuota.ts
/**
 * api.video
 * api.video is an API that encodes on the go to facilitate immediate playback, enhancing viewer streaming experiences across multiple devices and platforms. You can stream live or on-demand online videos within minutes.
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
* Deprecated
*/
export class AccountQuota {
    /**
    * Deprecated
    */
    'quotaUsed'?: number;
    /**
    * Deprecated
    */
    'quotaRemaining'?: number;
    /**
    * Deprecated
    */
    'quotaTotal'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "quotaUsed",
            "baseName": "quotaUsed",
            "type": "number",
            "format": ""
        },
        {
            "name": "quotaRemaining",
            "baseName": "quotaRemaining",
            "type": "number",
            "format": ""
        },
        {
            "name": "quotaTotal",
            "baseName": "quotaTotal",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AccountQuota.attributeTypeMap;
    }
    
    public constructor() {
    }
}