// models/BytesRange.ts
/**
 * api.video
 * api.video is an API that encodes on the go to facilitate immediate playback, enhancing viewer streaming experiences across multiple devices and platforms. You can stream live or on-demand online videos within minutes.
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export class BytesRange {
    /**
    * The starting point for the range of bytes for a chunk of a video.
    */
    'from'?: number;
    /**
    * The ending point for the range of bytes for a chunk of a video.
    */
    'to'?: number;
    /**
    * The total number of bytes in the provided range.
    */
    'total'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "from",
            "baseName": "from",
            "type": "number",
            "format": ""
        },
        {
            "name": "to",
            "baseName": "to",
            "type": "number",
            "format": ""
        },
        {
            "name": "total",
            "baseName": "total",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return BytesRange.attributeTypeMap;
    }
    
    public constructor() {
    }
}